Init18. TagController refactoring + new Routes + AdminOnly middleware

+++ backend/app/Http/Controllers/TagController.php
@@ -4,8 +4,11 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreTagRequest;
 use App\Http\Requests\UpdateTagRequest;
+use App\Http\Resources\PostResource;
+use App\Http\Resources\TagResource;
 use App\Models\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class TagController extends Controller
 {
@@ -15,7 +18,7 @@ class TagController extends Controller
     // Метод получения всех Тегов
     public function index()
     {
-        return Tag::all();
+        return TagResource::collection(Tag::all());
     }
 
     /**
@@ -24,11 +27,16 @@ class TagController extends Controller
     // Метод получения нового Тега
     public function store(StoreTagRequest $request)
     {
+        // Проверяем является ли Юзер Админом
+        if (!Auth::user()->is_admin) {
+            return response()->json(['message' => 'Forbidden'], 403);
+        }
+
         // Получаем экземпляр нового отвалидированного Тега
         $tag = Tag::create($request->validated());
 
-        // Возвращаем новый Тег в JSON со статусом 201
-        return response()->json($tag, 201);
+        // Возвращаем новый ресурс Тега
+        return new TagResource($tag);
     }
 
     /**
@@ -38,7 +46,7 @@ class TagController extends Controller
     // Метод получения конкретного Тега
     public function show(Tag $tag)
     {
-        return $tag;
+        return new TagResource($tag);
     }
 
     /**
@@ -47,11 +55,16 @@ class TagController extends Controller
     // Метод обновления Тега
     public function update(UpdateTagRequest $request, Tag $tag)
     {
+        // Проверяем является ли Юзер Админом
+        if (!Auth::user()->is_admin) {
+            return response()->json(['message' => 'Forbidden'], 403);
+        }
+
         // Обновляем текущий Тег валидированными данными
         $tag->update($request->validated());
 
-        // Возвращаем обновленный Тег в JSON
-        return response()->json($tag);
+        // Возвращаем обновленный Тег в виде ресурса
+        return new TagResource($tag);
     }
 
     /**
@@ -60,8 +73,21 @@ class TagController extends Controller
     // Метод удаления текущего Тега
     public function destroy(Tag $tag)
     {
+        // Проверяем является ли Юзер Админом
+        if (!Auth::user()->is_admin) {
+            return response()->json(['message' => 'Forbidden'], 403);
+        }
+
         $tag->delete();
 
         return response()->json(['message' => 'Tag deleted']);
+    }
+
+    // Метод получения всех Постов по Тегу
+    public function posts(Tag $tag)
+    {
+        $posts = $tag->posts()->with(['user', 'category'])->latest()->paginate(10);
+
+        return PostResource::collection($posts);
     }
 }
